+++ example/flake.nix
@@ -24,14 +24,10 @@
           in
           {
             cli = {
+              # environment.PC_DISABLE_TUI = true;
               # Global options for `process-compose`
-              global = {
+              options = {
                 no-server = true;
-              };
-              # Options for `process-compose up`
-              up = {
-                disable-dotenv = true;
-                theme = "Cobalt";
               };
             };
             settings = {