+++ CHANGELOG.md
@@ -3,8 +3,10 @@
 ## Unreleased
 
 - New features
-    - #81: Complete support for specifying process-compose CLI options (global and `up` subcommand arguments)
-      - **Breaking change**: Old options `httpServer` and `tui` were removed; users should use the new `cli` module to set all process-compose cli arguments and options.
+    - #81, #84: Support for specifying process-compose global CLI options
+      - **Breaking changes**:
+        - `preHook` and `postHook` are now inside `cli` module.
+        - Old options `httpServer` and `tui` were removed; users should use the new `cli` module to set all process-compose cli global options. TUI can be disabled using `cli.environment.PC_DISABLE_TUI = true;`
     - ~~#18: Add `testScript` option for adding flake checks based on nixosTest library.~~
     - #39: Allow `test` process to act as a test, which then gets run as part of flake checks.
     - #55: Add `lib` flake output - library of useful functions
@@ -12,7 +14,7 @@
     - New options
       - #52: Add `is_foreground` option
       - ~~#54: Add `apiServer` option to control REST API server~~
-      - $60: Add `httpServer.{enable, port, uds}` options to control the HTTP server.
+      - ~~$60: Add `httpServer.{enable, port, uds}` options to control the HTTP server.~~
       - #56: Add `preHook` and `postHook` for running commands before and after launching process-compose respectively.
       - #67: Add `ready_log_line`
       - #226: Add `availability.exit_on_skipped`